fix usage of atom-package-deps.install() API
every activation of the toolbar emits this warning message

```
Package-Deps] Unable to get loaded package 'tool-bar'
(anonymous) @ (...)/node_modules/atom-package-deps/lib/helpers.js:41
```

the reason is a wrong [API usage][8a28b864] of atom-package-deps.install().
the first parameter should be the name of the package which installs the
dependencies. the second argument is a boolean to prompt for confirmation
before install requested packages.

  [8a28b864]: https://github.com/steelbrain/package-deps#how-it-works